+++ octgnFX/Octgn/Launcher/HostGameSettings.xaml.cs
@@ -27,6 +27,7 @@ namespace Octgn.Launcher
         {
             InitializeComponent();
             _game = game;
+            textBox1.Text = Prefs.LastRoomName;
         }
 
         private void PageLoaded(object sender, RoutedEventArgs e)
@@ -67,6 +68,7 @@ namespace Octgn.Launcher
                 return;
             }
             if (_beginHost) return;
+            Prefs.LastRoomName = textBox1.Text;
             _startGameTimer = new Timer((object a)=> EndHostGame(-1) ,null,10000,Timeout.Infinite);
             progressBar1.Visibility = Visibility.Visible;
             button1.IsEnabled = false;