+++ octgnFX/Octgn/Launcher/HostedGameList.xaml.cs
@@ -1,6 +1,7 @@
 ﻿using System;
 using System.Linq;
 using System.Windows;
+using System.Windows.Controls;
 using System.Windows.Input;
 using Octgn.Controls;
 using Skylabs.Lobby;
@@ -13,19 +14,28 @@ namespace Octgn.Launcher
     /// </summary>
     public partial class HostedGameList
     {
+        private Border Refreshing;
         public HostedGameList()
         {
             InitializeComponent();
-            Program.LobbyClient.BeginGetGameList();
+            Refreshing = bRefreshing;
             Program.LobbyClient.OnDataRecieved += LobbyClientOnOnDataRecieved;
+            Program.LobbyClient.BeginGetGameList();
         }
 
         private void LobbyClientOnOnDataRecieved(object sender , Client.DataRecType type , object data)
         {
             if(type == Client.DataRecType.GameList)
                 ReloadGameList();
-            else if(type == Client.DataRecType.GamesNeedRefresh)
-                Program.LobbyClient.BeginGetGameList();
+            else if (type == Client.DataRecType.GamesNeedRefresh)
+            {
+                Dispatcher.Invoke(new Action(() =>
+                    {
+                    bRefreshing.Visibility = Visibility.Visible;
+                    Program.LobbyClient.BeginGetGameList();
+                    }));
+                
+            }
         }
 
         public event EventHandler OnGameClick;
@@ -33,34 +43,35 @@ namespace Octgn.Launcher
         private void ReloadGameList()
         {
             Dispatcher.Invoke(new Action(() =>
-                                             {
-                                                 stackPanel1.Children.Clear();
-                                                 var gids = new Guid[Program.GamesRepository.Games.Count];
-                                                 int count = 0;
-                                                 var filterList = new Guid[Program.GamesRepository.Games.Count];
-                                                 foreach (Data.Game game in Program.GamesRepository.Games)
-                                                 {
-                                                     gids[count] = game.Id;
-                                                     if (SimpleConfig.ReadValue("FilterGames_"+game.Name) == "False")
-                                                        filterList[count] = game.Id;
-                                                     count++;
-                                                 }
+                {
+                    bRefreshing.Visibility = Visibility.Collapsed;
+                    stackPanel1.Children.Clear();
+                    var gids = new Guid[Program.GamesRepository.Games.Count];
+                    int count = 0;
+                    var filterList = new Guid[Program.GamesRepository.Games.Count];
+                    foreach (Data.Game game in Program.GamesRepository.Games)
+                    {
+                        gids[count] = game.Id;
+                        if (!Prefs.getFilterGame(game.Name))
+                        filterList[count] = game.Id;
+                        count++;
+                    }
 
-                                                 HostedGameData[] gl =
-                                                     Program.LobbyClient.GetHostedGames().OrderByDescending(
-                                                         item => item.TimeStarted).ToArray();
-                                                 foreach (HostedGameData g in gl)
-                                                 {
-                                                     if (!gids.Contains(g.GameGuid) ||
-                                                         g.GameStatus != EHostedGame.StartedHosting) continue;
-                                                     var gs = new HostedGameListItem(g);
-                                                     if (g.GameStatus == EHostedGame.StartedHosting)
-                                                         gs.MouseUp += GsMouseUp;
-                                                     stackPanel1.Children.Add(gs);
-                                                     if (filterList.Contains(g.GameGuid))
-                                                        gs.Visibility= Visibility.Collapsed;
-                                                 }
-                                             }));
+                    HostedGameData[] gl =
+                        Program.LobbyClient.GetHostedGames().OrderByDescending(
+                            item => item.TimeStarted).ToArray();
+                    foreach (HostedGameData g in gl)
+                    {
+                        if (!gids.Contains(g.GameGuid) ||
+                            g.GameStatus != EHostedGame.StartedHosting) continue;
+                        var gs = new HostedGameListItem(g);
+                        if (g.GameStatus == EHostedGame.StartedHosting)
+                            gs.MouseUp += GsMouseUp;
+                        stackPanel1.Children.Add(gs);
+                        if (filterList.Contains(g.GameGuid))
+                        gs.Visibility= Visibility.Collapsed;
+                    }
+                }));
         }
 
         private void GsMouseUp(object sender, MouseButtonEventArgs e)
@@ -99,9 +110,16 @@ namespace Octgn.Launcher
         public void FilterGames(Guid g, Boolean show)
         {
             //let's cycle through the game list and hide/show them from the StackPanel
-            foreach (HostedGameListItem hg in stackPanel1.Children.Cast<HostedGameListItem>().Where(hg => hg.Game.GameGuid == g))
+            foreach(var hg in stackPanel1.Children)
             {
-                hg.Visibility = show ? Visibility.Visible : Visibility.Collapsed;
+                var h = hg as HostedGameListItem;
+                if(h != null)
+                {
+                    if(h.Game.GameGuid == g)
+                    {
+                        h.Visibility = show ? Visibility.Visible : Visibility.Collapsed;
+                    }
+                }
             }
         }
     }